+++ NEWS.md
@@ -28,6 +28,7 @@ Red Panda C++ Version 2.23
   - fix: GLFW project template.
   - fix: Inherited class/struct members are not correctly shown in the completion suggestions.
   - enhancement: Sort symbols by their declaration pos in the Class Browser, if not sort by alpha order.
+  - fix: Keyword asm is not correctly parsed.
 
 Red Panda C++ Version 2.22